requirements.txt
Django==1.8.4
django-cors-headers==1.1.0
django-extensions==1.5.7
django-filter==0.11.0
djangorestframework ==3.2.3